+++ examples/mf6/hondsrug.py
@@ -21,6 +21,8 @@ In overview, the model features:
 # %% Import packages
 # We'll start with the usual imports, and an import from scipy.
 
+from copy import deepcopy
+
 import matplotlib.pyplot as plt
 import numpy as np
 import pandas as pd
@@ -612,6 +614,9 @@ simulation.create_time_discretization(
 #   describes how to add it to yours.
 
 modeldir = imod.util.temporary_directory()
+mask = deepcopy(simulation["GWF_1"].domain)
+mask.values[:,2:8,2:8] = 0
+simulation["GWF_1"].mask_all_packages(mask)
 simulation.write(modeldir, binary=False)
 simulation.run()
